refactor: split index.ts functionalities and create index-like entrypoint
